requirements.txt
# Core orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0

# LangSmith observability
langsmith>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# GitHub API
PyGithub>=2.3.0

# RAG & Vector DB
qdrant-client>=1.9.0
langchain-qdrant>=0.1.0
google-generativeai>=0.7.0  # for Gemini embeddings

# RAG Evaluation — pinned: ragas 0.4.x private API used in evaluation/ragas_eval.py
ragas==0.4.3
datasets>=2.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Persistent checkpointing for HITL resilience
langgraph-checkpoint-sqlite>=3.0.0
aiosqlite>=0.20.0   # required for AsyncSqliteSaver

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
httpx>=0.27.0
rich>=13.0.0   # pretty console output
